animation moved from App.jsx, some react components added, page structure updated to fit window, tab name and icon added, misc css and images added
