+++ src/App.jsx
@@ -1,4 +1,4 @@
-import { useState } from 'react'
+// import { useState } from 'react'
 import NavBar from './components/navBar';
 import LeftMainContainer from './components/MainContainers/LeftMainContainer'; 
 import RightMainContainer from './components/MainContainers/RightMainContainer'; 
@@ -6,18 +6,6 @@ import RightMainContainer from './components/MainContainers/RightMainContainer';
 import ArrayContainer from './components/ArrayContainer';
 
 function App() {
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const arrayData = [1, 2, 3, 4]; 
-  const handleButtonClick = () => {
-    setCurrentIndex((prevIndex) => {
-      if (prevIndex < arrayData.length - 1) {
-        return prevIndex + 1;
-      } else {
-        return 0;
-      }
-    });
-  };
-
   return (
     <>
      <div>
@@ -27,10 +15,6 @@ function App() {
         <RightMainContainer/>
       </div>
 
-
-
-      <button onClick={handleButtonClick}>Move</button>
-      <ArrayContainer arrayData={arrayData} currentIndex={currentIndex} />
     </div>
     </>
   )
