+++ src/components/MainContainers/RightMainContainer.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-
+import Animation from '../animation';
 
 
 const RightMainContainer = () => {
     return (
-      <div>
+      <div className="right">
         <div className="right-main-container">
-
+            <Animation/>
         </div>
       </div>
     );
